expense_service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-jose
python-dotenv
pydantic
requests
cryptography